multisite-sitepaths: three class solution
